Went from a barely working web app, to something.
Fixed the main function JavaScript that sends the requests.
Added const variables with useState() for the text input.
Fixed the button.
Added a results part of the page.

What does the app do now? It sends a request to a Zillow API once the button is pushed.

The results shown on screen are JSON objects with results of property listings in a given area.